delete old feedback component
