+++ src/app/context/FeedbackContext.jsx
@@ -4,8 +4,9 @@ import { useFeedbackBox } from '@nypl/design-system-react-components';
 import { trackDiscovery } from '../utils/utils';
 
 export const FeedbackBoxContext = createContext(null);
-export const FeedbackBoxProvider = ({ children }) => {
-  const [itemMetadata, setItemMetadata] = useState(null)
+export const FeedbackBoxProvider = ({ children, value }) => {
+  console.log(value)
+  const [itemMetadata, setItemMetadata] = useState(value && value.itemMetadata ? value.itemMetadata : null)
   const { FeedbackBox, isOpen, onOpen, onClose } = useFeedbackBox()
   const openFeedbackBox = () => {
     trackDiscovery('Feedback', 'Open')
