--- src/client/styles/components/Feedback.scss
@@ -1,131 +1,0 @@
-.feedback {
-  bottom: 0;
-  right: 0;
-  z-index: 100;
-  text-align: right;
-  background: $white;
-  color: $black;
-  position: fixed;
-  background: inherit;
-  max-width: 571px;
-  min-width: 571px;
-
-  button.feedback-button {
-    background: $researchPrimary;
-    border: 1px solid $researchPrimary;
-    box-sizing: border-box;
-    color: $white;
-    display: inline-block;
-    font-size: 1rem;
-    padding: .5rem 1rem;
-    position: relative;
-    text-align: center;
-  
-    &:hover {
-      background: $white;
-      color: $researchPrimary;
-    }
-  }
-}
-
-
-.cancel-button {
-  border: 1px solid $borderDefault;
-  font-size: 1rem;
-  margin-right: 15px;
-  text-align: center;
-}
-
-
-#feedback-menu {
-  display: none;
-  background: $white;
-  border: 1px solid $borderDefault;
-  box-shadow: -1px 0 1px 0 rgba(0,0,0,0.3);
-  padding: 50px 42px;
-  text-align: left;
-  height: 472px;
-
-  &.active {
-    display: block;
-  }
-
-  label {
-    span {
-      font-size: .75rem;
-      font-weight: 300;
-    }
-  }
-
-  p {
-    margin: 107px 40px;
-    font-size: 20px;
-    font-weight: 500;
-  }
-
-  input,
-  select,
-  textarea {
-    display: block;
-    width: 100%;
-    background-color: white;
-    font-size: 1rem;
-    line-height: 1.2;
-    box-sizing: border-box;
-    border: 1px solid $grayMedium;
-  }
-
-  textarea {
-    height: 8em;
-  }
-
-  button {
-    font-size: 1rem;
-    display: inline;
-    width: 164px;
-  }
-
-  iframe {
-    display: none;
-  }
-
-  .helper-text {
-    height: 21px;
-  }
-}
-
-.privacy-policy {
-  display: flow-root;
-  a {
-    float: right;
-    font-size: .75rem;
-  }
-}
-
-@media (max-width: $nypl-breakpoint-large) {
-  #feedback-menu {
-    height: 400px;
-    padding: 34px;
-
-    h1 {
-      font-size: 1rem;
-      margin-bottom: 7px;
-    }
-
-    label {
-      margin-bottom: 4px;
-    }
-
-    textarea {
-      height: 6em;
-    }
-
-    button {
-      margin-top: 13px;
-      width: 98px;
-      height: 25px;
-      font-size: .875rem;
-      padding: 0;
-    }
-  }
-}
